--- src/main/java/kr/hhplus/be/server/product/controller/docs/ProductSchemaDescription.java
@@ -1,10 +1,0 @@
-package kr.hhplus.be.server.product.controller.docs;
-
-public interface ProductSchemaDescription {
-
-    String id = "상품 ID";
-    String name = "상품 이름";
-    String price = "상품 가격";
-    String quantity = "상품 수량";
-
-}
