--- src/main/java/kr/hhplus/be/server/wallet/controller/docs/WalletSchemaDescription.java
@@ -1,9 +1,0 @@
-package kr.hhplus.be.server.wallet.controller.docs;
-
-public interface WalletSchemaDescription {
-
-    String userId = "사용자 ID";
-    String amount = "금액";
-    String balance = "잔액";
-
-}
