+++ src/main/java/kr/hhplus/be/server/wallet/presentation/docs/WalletSchemaDescription.java
@@ -1,0 +1,9 @@
+package kr.hhplus.be.server.wallet.presentation.docs;
+
+public interface WalletSchemaDescription {
+
+    String userId = "사용자 ID";
+    String amount = "금액";
+    String balance = "잔액";
+
+}
